requirements.txt
paho-mqtt
requests
RPi.GPIO